--- components/App.jsx
@@ -1,19 +1,0 @@
-import React, { Component } from 'react';
-import AddListing from './AddListing.jsx';
-
-class App extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-
-    };
-  }
-  render() {
-    return (<div> APP.JSX
-      <AddListing />
-    </div>);
-  }
-}
-
-export default App;
