+++ package.json
@@ -20,7 +20,6 @@
     "babel-preset-react-hmre": "^1.1.1",
     "babel-register": "^6.9.0",
     "body-parser": "^1.15.1",
-<<<<<<< 025da1398e4f77304ce5d4a98f5de72c3330797b
     "eslint": "^2.11.1",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.8.1",
@@ -29,27 +28,15 @@
     "express": "^4.13.4",
     "file-loader": "^0.8.5",
     "mongoose": "^4.4.20",
-=======
-    "bootstrap": "^3.3.6",
-    "eslint": "^2.10.2",
-    "express": "^4.13.4",
-    "express-session": "^1.13.0",
-    "jquery": "^2.2.4",
->>>>>>> Refactor back end, move files around into public, add linting and webpack
     "multer": "^1.1.0",
     "nodemon": "^1.9.2",
     "q": "^1.4.1",
     "react": "^15.1.0",
     "react-dom": "^15.1.0",
-<<<<<<< e56f428dc453dd8c642fe0cf2a5e7ea5ea239caf
-<<<<<<< 025da1398e4f77304ce5d4a98f5de72c3330797b
     "react-dropzone": "^3.5.0",
     "superagent": "^2.0.0",
     "webpack": "^1.13.1",
-=======
-=======
     "react-redux": "^4.4.5",
->>>>>>> Get Redux, react, react-router set up
     "react-router": "^2.4.1",
     "react-router-redux": "^4.0.5",
     "redux": "^3.5.2",
@@ -71,16 +58,10 @@
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.5.0",
-    "eslint": "^2.11.1",
-    "eslint-config-airbnb": "^9.0.1",
     "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-jsx-a11y": "^1.2.2",
-    "eslint-plugin-react": "^5.1.1",
     "webpack": "^1.13.1",
     "morgan": "^1.7.0",
     "nodemon": "^1.9.2",
->>>>>>> Refactor back end, move files around into public, add linting and webpack
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.10.0"
   }
