+++ webpack.config.js
@@ -16,13 +16,6 @@ module.exports = {
     new webpack.NoErrorsPlugin(),
   ],
   module: {
-    // preLoaders: [
-    //   {
-    //     test: /\.jsx?$/,
-    //     loader: 'eslint',
-    //     exclude: /(node_modules|bower_components)/,
-    //   },
-    // ],
     loaders: [
       {
         test: /\.jsx?$/,
